requirements.txt
# System: latexmk, pdflatex (texlive)
PyMuPDF>=1.23.0
python-pptx>=1.0.0
